Creating shared app test fixture - #618
