+++ test/conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = ["test.helpers.auth", "test.helpers.database", "test.helpers.request"]
+pytest_plugins = ["test.helpers", "test.helpers.auth", "test.helpers.database", "test.helpers.request"]
